cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click",
    "snowflake-connector-python[pandas]",
    "toml>=0.10.2",
    "sqlfluff",
    "sqllineage",
]
